requirements.txt
# PACBench Requirements

# Core dependencies
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# For summary table generation
tabulate>=0.9.0



